Skeleton rom; Combined POC and Credits into ROM
Took the rng_map proof of concept and the credits screen and put them
in a single ROM that uses MMC1.  The mapper is setup for three 16k PRG
banks and one fixed 16k bank with two 8k CHR banks and 8k PRG RAM.

States are implemented, but it needs some work.

Using the cc65 toolkit with a proper Makefile.  A new label conversion
script is needed, however.