Rearranged tiles in CHR to make room for BG art
All the text has been moved to CHR0 (the "sprite" bank) leaving CHR0
mostly empty for background art.  A new constant has been added
(PPU_CTRL_TITLE) to set the correct CHR banks for the title, game over,
and high scores screens as well as the status bar in the game screen.

The background tile used for the sprite zero hit has been moved to $0F
to avoid collision with the title screen's cursor.  This tile is on
both CHR0 and CHR1.

Tile $0F is also used on the high scores screen as the black background
tile.  The background tiles for the high scores have also been moved to
CHR0, as CHR bank switching will not be feasible for text and
background in separate banks.

Added the SPRITE_TILE_START constant for the seed screen for the offset
at which the large hexadecimal metasprite tiles start.

Lastliy, added some notes to dot-plan.txt about backgrounds and themes.