Decouple gen and buffer; Buffer PPU address
Decouple buffering meta columns from their generation.  Meta columns
will be buffered when they are needed.  This allows for a double-wide
obstacle to be generated and then drawn on the next two frames.

The start PPU address is buffered in RAM so that calculation doesn't
happen during NMI.  Also fixed an off-by-one error with the nametable
calculation causing the first column on the second nametable to draw
over top of the first column on the first nametable.

Added and clarified a bunch of comments.