Rework collision detection to work with pits
Cache the two meta columns that the player is occupying and check
collision against the cache.  Collision is checked in layers, starting
with the bottom meta tile and working up depending on the current Y
position of the player meta sprite.

This has the added benefit of being able to use half height walls with
no additional code.