Update credits code
This brings the credits screen up to date with the current development
rom.  Tier 2 and 3 names almost work correctly (with the exception of
the last name of the first nametable).

The only differences between the code here and the code in the dev rom
is to make it work (lol).  Namely the frame and NMI pointers.