Implement background drawing
This starts implementing drawing a background on the playfield that is
above the randomly generated map.  It uses OP codes similar to the
credits screen, but with some changes.  Available OP codes are EOD,
ByteList, RLE, Repeating RLE and Attribute.

Attribute stuff still needs to be reworked a little bit.

There is currently only enough data for a single background building
(two columns).