Start using different palettes for the foreground
The top two meta tiles of the game field use the third palette, the
bottom two meta tiles use the second palette, and the status bar uses
the first palette.

Unchanged is the background which still uses the fourth palette.

Also added a new meta tile for drawing a blank background on bottom
half of the play field (used with pits).