Add the background transition tile row
Due to this row not lining up with the attribute table nicely, we are
using the last color on the background palette as the "transition"
color (green, in this case).

This row is defined by a single byte at the end of the background data
that designates the first tile of a four tile sequence that is repeated
at the bottom of the background.