Put a skyline on the title screen
Removed some magic numbers in the background drawing routine and
replaced them with variables in zero page.  This allows the skyline to
be drawn at different heights on the screen.

Also some minor fixes for the title screen to fit the skyline, as well
as adding a sprite zero to swap background CHR data.