Further optimize the credits storage format
(taken from the credits dev repo)

The format for names is now one byte for the name op code, one byte for
both the attribute data and length of the name, then a list of bytes
for the name.

The math for centering names is now done in assembly instead of
beforehand in the generation script.