Simplify gamestate code with a lookup table
Replace a bunch of comparisons and branches with a lookup table of init
entry points.  The mapper is set to a known state, then the address for
the init code of the new state is loaded from the lookup table and
jumped to.

Gamestate values have been changed to plain integers in a enum to be
used as the index for the lookup table.