Remove ZP0 line and CHR swap
Removed the collision line for sprite zero and moved sprite zero up a
scanline and to the right side of the screen.  This triggers the scroll
to take place during the horizontal blank period instead of during the
draw of a scanline.

Move the status bar tiles to the second CHR pattern table and don't
swap banks or swap background pattern tables with the sprite 0 hit.
This removes the need for converting to ASCII for the status bar.