Fix high scores screen
Fixed the high scores screen to use hex values instead of ASCII when
drawing the seed and score to the screen.  This also changes how the
seeds are stored in PRG RAM from ascii to binary (four bytes -> two
bytes).

Also made a minor change in game.asm that swapped a CLC / ADC with an
ORA operation.